+++ HISTORY.md
@@ -14,7 +14,7 @@ All notable bugs/errors to this project will be documented in this file.
 
 2. [On 16 Mar 2024 checkov dynamoDB autoscaling issue and resolved](./screenshots/16032024-checkov-autoScaling_issue.png)
 
-3. [On 17 Mar 2024 checkov s3 bucket versioning issue and resouce block added](./screenshots/17032024-checkov-s3_bucket_versioning_issue.png)
+3. [On 21 Mar 2024 checkov s3 bucket versioning issue and resouce block added](./screenshots/21032024-checkov-s3_bucket_versioning_issue.png)
 
 4. [On 17 Mar 2024 checkov ensure access control lists for S3 buckets are disabled](./screenshots/17032024-checkov-s3_bucket_ACL_issue.png)